pyproject.toml
[tool.poetry]
name = "reverse-proxy"
version = "0.1.0"
description = ""
authors = ["Vadym Stupakov <vadim.stupakov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.2"
flask-socketio = "^5.3.6"
requests = "^2.31.0"
python-socketio = {extras = ["client"], version = "^5.11.1"}
expression = {extras = ["pydantic"], version = "^5.0.2"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
